# Lazily create XmlPath in ValidationErrorInfo