If any errors are encountered, `ValidationErrorInfo` automatically creates an `Xmlpath` instance. However, this is somewhat expensive to create and may not be used. The validation perf test on master is this:

| Method     | Mean     | Error   | StdDev  | Gen 0     | Allocated |
| ---------- | -------- | ------- | ------- | --------- | --------- |
| Before     | 141.8 ms | 2.75 ms | 2.70 ms | 5000.0000 | 23.72 MB  |
| After      | 138.5 ms | 4.55 ms | 4.68 ms | 4000.0000 | 19.87 MB  |

There's a slight reduction in time, but a large reduction in Gen0 and overall allocations (~17%)